Add DISPLAY_FORMAT
Choose which content to display in the notification. This will display
them in the order that they are placed in the list. Possible options are:
  app_name
  notification_id
  app_icon,
  summary
  body

This can also take raw strings that will be placed in every notification.
(Good for parsing the output with key phrases delineating types of
content)

DISPLAY_FORMAT = ['APPN', 'app_name', 'SUMM', 'summary', 'BODY', 'body']
DISPLAY_FORMAT = ['app_name', 'summary', 'body']